requirements.txt
transifex-client
django-cms>=2.3
django
